pyproject.toml
[project]
name = "notify"
version = "0.2.1"
description = "some of predifined notify method to various destination eg. discord, email"
readme = "README.md"
authors = [
    { name = "Phanuwat Suwan", email = "phanuwatsuwan04@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "apache-airflow==2.10.5",
    "discord-webhook>=1.4.1",
    "dotenv>=0.9.9",
    "pendulum>=3.2.0",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
